AS-2023-2024/5X-TPSIT/es-07_2023-12-18_MovieList/requirements.txt
uvicorn
httptools
fastapi
jinja2
mysql-connector-python